Remove PersistencePreview and add base CalorieEntryView
