--- QuickCal/Previews/PersistencePreview.swift
@@ -1,31 +1,0 @@
-//
-//  PersistencePreview.swift
-//  QuickCal
-//
-//  Created by Will Boland on 9/12/22.
-//
-
-import Foundation
-
-extension PersistenceController {
-    static var preview: PersistenceController = {
-        let result = PersistenceController(inMemory: true)
-        let viewContext = result.container.viewContext
-        for index in 0..<10 {
-            let newItem = CalorieEntry(context: viewContext)
-            newItem.title = "Calorie \(index)"
-            newItem.amount = 100
-            newItem.timestamp = Date()
-        }
-        do {
-            try viewContext.save()
-        } catch {
-            // Replace this implementation with code to handle the error appropriately.
-            // fatalError() causes the application to generate a crash log and terminate.
-            // You should not use this function in a shipping application, although it may be useful during development.
-            let nsError = error as NSError
-            fatalError("Unresolved error \(nsError), \(nsError.userInfo)")
-        }
-        return result
-    }()
-}
